Refactoring renderHtmlWithRouterLinks func into global utils
-Moving code to staticHtmlUtils.js so it can be imported into any component that needs it
-Adding better commenting/docs
-Removing function from SolutionsTemplate and importing from utils
